+++ AssetStudio/SerializedFileHeader.cs
@@ -8,9 +8,9 @@ namespace AssetStudio
     public class SerializedFileHeader
     {
         public uint m_MetadataSize;
-        public uint m_FileSize;
+        public long m_FileSize;
         public uint m_Version;
-        public uint m_DataOffset;
+        public long m_DataOffset;
         public byte m_Endianess;
         public byte[] m_Reserved;
     }
